BREAKING(spot): removing spot scheduler
Previously spot instances could be only
terminated, That why the spot instance
scheduler terminated the instances and
not only shutdown. AWS has now added a
new category of spot instance (persistent
request) which can be stopped and started.

This commit remove the spot instance
scheduler which is now merged in the
instance scheduler.