fix(lambda): removing source_code_hash param
When source_code_hash is defined in aws_lambda_function
Terraform will check the hash that the zip which contains
the code used by aws lambda. This value can changed
depending of your operating system even if the code has
not changed. This may be a problem when Terraform is
executed on different operating systems.

This commit remove the aws lambda param source_code_hash
in the Terraform code.

https://github.com/diodonfrost/terraform-aws-lambda-scheduler-stop-start/pull/11