fix(lambda): add git version in the filename used by lambda
Previously, the filename used by the lambda
function was always the same, that was problematic
if the source code changed the lambda function won't
be changed because we no longer use the parameter
aws source_code_hash.

This commit fix that by changing the filename according
to the latest git tag.