feat(terraform): add redshift scheduler
This commit introduces a new feature to our Terraform configuration, enabling the addition of a Redshift scheduler module. This module allows for the automated stopping or starting of Redshift clusters based on predefined schedules. By leveraging this scheduler, we can effectively manage the usage and costs associated with our Redshift clusters, ensuring they are active only when needed and maximizing resource efficiency.