fix(lambda): use output_base64sha256 directly instead of function filebase64sha256
"source_code_hash = filebase64sha256(data.archive_file.this.output_path)"
can be called before the "data.archive_file.this" resource has created
the zip file and may fails because the file does not exist.

The cause of the issue here is calling a function with a known argument,
the workaround would be to directly use the output_base64sha256
attribute of the data source instead of calling filebase64sha256.

this commit replaces the "filebase64sha256" function by the "output_base64sha256"

Thanks to @pvillaverde for pointing this out!